+++ features/step_definitions/marketplace/mp_user_group_affiliations_steps.rb
@@ -8,6 +8,11 @@ Given(/^I select to verify the Military Group Affiliation$/) do
   @marketplace_group_affiliations_page.click_verify_military_affiliation_link
 end
 
+
+Given(/^I select to verify the Teacher Group Affiliation$/) do
+  @marketplace_group_affiliations_page.click_verify_teacher_affiliation_link
+end 
+
 Given(/^I select to verify the First Responder Group Affiliation$/) do
   @marketplace_group_affiliations_page.click_verify_first_responder_affiliation_link
 end