+++ features/support/app_pages/marketplace/mp_group_affiliations_page.rb
@@ -15,9 +15,14 @@ class MarketplaceGroupAffiliationsPage < IDmeBase
 
 	def click_verify_military_affiliation_link
 		find(:link, :text =>"Military").click
-	end 
+	end
+
+	def click_verify_teacher_affiliation_link
+		find(:link, :text =>"Teachers").click 
+	end  
 
 	def click_verify_first_responder_affiliation_link
 		find(:link, :text =>"First Responders").click 
 	end 
+
 end